Added the ability to nuke all comments of a post in comment-nuke app (#157)
## 💸 TL;DR

Added the ability to remove & lock all comments of a post by introducing
the `handleNukePost` function.

`handleNukePost` allows a moderator to lock or remove all comments in a
post. It fetches a post using the provided post ID, checks if the
current user has the necessary mod permissions, and then proceeds to
lock or remove all comments in the post based on the provided
properties. If the `skipDistinguished` property is true, distinguished
comments are skipped. The function logs any failed attempts to lock or
remove comments and returns a success message along with the number of
comments handled.

Also added a separate form - `nukePostForm` in main.ts for this purpose.

Created this PR after discussing with @pl00h on Discord, the old PR #5
was closed due to changes in the repository structure.

## ✅ Checks
<!-- Make sure your pr passes the CI checks and do check the following
fields as needed - -->
- [x] CI tests (if present) are passing
- [x] Adheres to code style for repo
- [x] Contributor License Agreement (CLA) completed if not a Reddit
employee

---------

Signed-off-by: Nisarga <git@nisarga.me>