Comment Mop: Fix two bugs in app relating to comments that should be removed staying visible (#170)
<!-- If this pull request closes an issue, please mention the issue number below -->
Closes #168 
Closes #169

## 💸 TL;DR
In the latest release of Comment Mop, two regressions were introduced as described in the above two issues. The logic used to find comments to remove was improved to be more efficient, but this changed the behaviour compared to the original version.

## 🧪 Testing Steps / Validation
Create comment hierarchies on a post as described in the two issues.

Run Comment Mop against the top level comment in the hierarchy, once with "Ignore Distinguished" turned off, and again with it turned on.

Validate that the effect of mopping the thread is correct:

For #168, we should observe that the top level comment where Mop was run is correctly removed. For #169, we should observe that the mop continues below the distinguished comment.

Repeat for the post-level mop (only #169 needs to be tested this time). Observe that any comments beneath a distinguished comment are removed, but the distinguished comment is retained when "Ignore distinguished comments" is selected.

## ✅ Checks
<!-- Make sure your pr passes the CI checks and do check the following fields as needed - -->
- [x] CI tests (if present) are passing
- [x] Adheres to code style for repo
- [x] Contributor License Agreement (CLA) completed if not a Reddit employee