feat: Allow explicit definition of new version when uploading app (#231)
## 💸 TL;DR

I want to specify the version of my app when uploading.

## 📜 Details

Not a breaking change.

## 🧪 Testing Steps / Validation

I'm unable to lint, test etc. because it isn't possible to install all
of the dependencies, as they depend on access to `snooguts.net`. The
change is small enough that it should, hopefully, in theory, work, and
be easy enough for someone to test.

## ✅ Checks

- [ ] CI tests (if present) are passing
- [ ] Adheres to code style for repo
- [x] Contributor License Agreement (CLA) completed if not a Reddit
employee